@@ -39,7 +39,8 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": "src/",
+            "Symfony\\Bundle\\FrameworkBundle\\Test\\": "vendor/symfony/framework-bundle/Test/"
         }
     },
     "autoload-dev": {
@@ -74,6 +75,9 @@
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.5",
-        "symfony/maker-bundle": "^1.57"
+        "phpunit/phpunit": "11",
+        "symfony/browser-kit": "7.0.*",
+        "symfony/maker-bundle": "^1.57",
+        "symfony/phpunit-bridge": "^7.1"
     }
 }
